# *: batch process peer task