--- a/pkg/cluster/cluster.go
+++ b/pkg/cluster/cluster.go
@@ -35,9 +35,7 @@ type Cluster interface {
 func HandleStatsAsync(c Cluster, region *core.RegionInfo) {
 	c.GetHotStat().CheckWriteAsync(statistics.NewCheckExpiredItemTask(region))
 	c.GetHotStat().CheckReadAsync(statistics.NewCheckExpiredItemTask(region))
-	reportInterval := region.GetInterval()
-	interval := reportInterval.GetEndTimestamp() - reportInterval.GetStartTimestamp()
-	c.GetHotStat().CheckWriteAsync(statistics.NewCheckPeerTask(region, region.GetPeers(), region.GetWriteLoads(), interval))
+	c.GetHotStat().CheckWriteAsync(statistics.NewCheckWritePeerTask(region))
 	c.GetCoordinator().GetSchedulersController().CheckTransferWitnessLeader(region)
 }