--- a/pkg/statistics/utils/kind.go
+++ b/pkg/statistics/utils/kind.go
@@ -14,10 +14,6 @@
 
 package utils
 
-import (
-	"github.com/pingcap/kvproto/pkg/metapb"
-)
-
 const (
 	// BytePriority indicates hot-region-scheduler prefer byte dim
 	BytePriority = "byte"
@@ -230,8 +226,8 @@ func (rw RWType) DefaultAntiCount() int {
 	}
 }
 
-// GetLoadRatesFromPeer gets the load rates of the read or write type from PeerInfo.
-func (rw RWType) GetLoadRatesFromPeer(peer *metapb.Peer, deltaLoads []float64, interval uint64) []float64 {
+// GetLoadRates gets the load rates of the read or write type.
+func (rw RWType) GetLoadRates(deltaLoads []float64, interval uint64) []float64 {
 	loads := make([]float64, DimLen)
 	for dim, k := range rw.RegionStats() {
 		loads[dim] = deltaLoads[k] / float64(interval)