--- a/server/cluster/cluster.go
+++ b/server/cluster/cluster.go
@@ -959,8 +959,7 @@ func (c *RaftCluster) HandleStoreHeartbeat(heartbeat *pdpb.StoreHeartbeatRequest
 				utils.RegionWriteKeys:     0,
 				utils.RegionWriteQueryNum: 0,
 			}
-			peerInfo := core.NewPeerInfo(peer, loads, interval)
-			c.hotStat.CheckReadAsync(statistics.NewCheckPeerTask(peerInfo, region))
+			c.hotStat.CheckReadAsync(statistics.NewCheckPeerTask(region, loads))
 		}
 	}
 	for _, stat := range stats.GetSnapshotStats() {