… pkg/statistics/hot_peer_cache_test.go
… server/cluster/cluster_test.go
… tools/pd-ctl/tests/hot/hot_test.go