<!--

Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?


<!--

Please create an issue first to describe the problem.
There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Ref #7897.

### What is changed and how does it work?

```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/tikv/pd/server/cluster
cpu: Intel(R) Xeon(R) Gold 6240 CPU @ 2.60GHz
                    │    old.txt    │               new.txt                │
                    │    sec/op     │    sec/op     vs base                │
HandleStatsAsync-72   610.2n ± 171%   370.1n ± 27%  -39.35% (p=0.000 n=10)

                    │   old.txt   │              new.txt               │
                    │    B/op     │    B/op     vs base                │
HandleStatsAsync-72   128.00 ± 3%   24.00 ± 0%  -81.25% (p=0.000 n=10)

                    │  old.txt   │              new.txt               │
                    │ allocs/op  │ allocs/op   vs base                │
HandleStatsAsync-72   5.000 ± 0%   3.000 ± 0%  -40.00% (p=0.000 n=10)
```

<!--

You could use the "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

```commit-message
```

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of these tests must be included. -->

- Unit test

### Release note

<!--

A bugfix or a new feature needs a release note. If there is no need to give a release note, just leave it with the `None`.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

-->

```release-note
None.
```